standardize route documentation
