+++ API/Controllers/ClientsController.cs
@@ -1,3 +1,4 @@
+using System.ComponentModel.DataAnnotations;
 using System.Diagnostics.CodeAnalysis;
 using API.Authorization;
 using API.DTOs;
@@ -16,16 +17,16 @@ public class ClientsController(IOAuthClientService clientService) : Controller
     /// <summary>
     /// Set the rate limit for a client
     /// </summary>
-    /// <param name="id">The client id</param>
+    /// <param name="id">Client id</param>
     /// <param name="rateLimitOverride">The new rate limit for the client</param>
-    /// <response code="404">If the provided id does not belong to a client</response>
-    /// <response code="200">Returns the patched client</response>
+    /// <response code="404">A client matching the given id does not exist</response>
+    /// <response code="200">Returns the updated client</response>
+    [HttpPost("{id:int}/ratelimit")]
     [Authorize(Roles = OtrClaims.Roles.Admin)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType<OAuthClientDTO>(StatusCodes.Status200OK)]
     [SuppressMessage("ReSharper", "StringLiteralTypo")]
-    [HttpPost("{id:int}/ratelimit")]
-    public async Task<IActionResult> PatchRateLimitAsync(int id, int rateLimitOverride)
+    public async Task<IActionResult> PatchRateLimitAsync(int id, [FromBody][Required] int rateLimitOverride)
     {
         OAuthClientDTO? client = await clientService.GetAsync(id);
