+++ API/Controllers/GameScoresController.Admin.cs
@@ -1,3 +1,4 @@
+using System.ComponentModel.DataAnnotations;
 using API.Authorization;
 using API.DTOs;
 using API.Utilities.Extensions;
@@ -13,15 +14,15 @@ public partial class GameScoresController
     /// Creates an admin note for a score
     /// </summary>
     /// <param name="id">Score id</param>
-    /// <response code="404">If a score matching the given id does not exist</response>
-    /// <response code="400">If the authorized user does not exist</response>
+    /// <response code="404">A score matching the given id does not exist</response>
+    /// <response code="400">The authorized user does not exist</response>
     /// <response code="200">Returns the created admin note</response>
     [HttpPost("{id:int}/notes")]
-    [Authorize(Roles = $"{OtrClaims.Roles.Admin}")]
+    [Authorize(Roles = OtrClaims.Roles.Admin)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
     [ProducesResponseType<AdminNoteDTO>(StatusCodes.Status200OK)]
-    public async Task<IActionResult> CreateAdminNoteAsync(int id, [FromBody] string note)
+    public async Task<IActionResult> CreateAdminNoteAsync(int id, [FromBody][Required] string note)
     {
         if (!await gameScoresService.ExistsAsync(id))
         {
@@ -35,42 +36,24 @@ public partial class GameScoresController
     }
 
     /// <summary>
-    /// List all admin notes for a score
-    /// </summary>
-    /// <param name="id">Score id</param>
-    /// <response code="404">If a score matching the given id does not exist</response>
-    /// <response code="200">Returns all admin notes from a score</response>
-    [HttpGet("{id:int}/notes")]
-    [Authorize(Roles = $"{OtrClaims.Roles.Admin}")]
-    [ProducesResponseType(StatusCodes.Status404NotFound)]
-    [ProducesResponseType<IEnumerable<AdminNoteDTO>>(StatusCodes.Status200OK)]
-    public async Task<IActionResult> ListAdminNotesAsync(int id)
-    {
-        if (!await gameScoresService.ExistsAsync(id))
-        {
-            return NotFound();
-        }
-
-        return Ok(await adminNoteService.ListAsync<GameScoreAdminNote>(id));
-    }
-
-    /// <summary>
     /// Updates an admin note for a score
     /// </summary>
     /// <param name="id">Score id</param>
     /// <param name="noteId">Admin note id</param>
     /// <response code="404">
-    /// If a score matching the given id does not exist.
-    /// If an admin note matching the given noteId does not exist
+    /// A score matching the given id does not exist
+    /// or an admin note matching the given noteId does not exist
     /// </response>
-    /// <response code="403">If the requester did not create the admin note</response>
     /// <response code="200">Returns the updated admin note</response>
     [HttpPatch("{id:int}/notes/{noteId:int}")]
-    [Authorize(Roles = $"{OtrClaims.Roles.Admin}")]
+    [Authorize(Roles = OtrClaims.Roles.Admin)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
-    [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [ProducesResponseType<AdminNoteDTO>(StatusCodes.Status200OK)]
-    public async Task<IActionResult> UpdateAdminNoteAsync(int id, int noteId, [FromBody] string note)
+    public async Task<IActionResult> UpdateAdminNoteAsync(
+        int id,
+        int noteId,
+        [FromBody][Required] string note
+    )
     {
         if (!await gameScoresService.ExistsAsync(id))
         {
@@ -97,14 +80,13 @@ public partial class GameScoresController
     /// <param name="id">Score id</param>
     /// <param name="noteId">Admin note id</param>
     /// <response code="404">
-    /// If a score matching the given id does not exist.
-    /// If an admin note matching the given noteId does not exist
+    /// A score matching the given id does not exist
+    /// or an admin note matching the given noteId does not exist
     /// </response>
     /// <response code="200">Returns the updated admin note</response>
     [HttpDelete("{id:int}/notes/{noteId:int}")]
-    [Authorize(Roles = $"{OtrClaims.Roles.Admin}")]
+    [Authorize(Roles = OtrClaims.Roles.Admin)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
-    [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [ProducesResponseType<AdminNoteDTO>(StatusCodes.Status200OK)]
     public async Task<IActionResult> DeleteAdminNoteAsync(int id, int noteId)
     {
