+++ API/Controllers/GamesController.Admin.cs
@@ -1,3 +1,4 @@
+using System.ComponentModel.DataAnnotations;
 using API.Authorization;
 using API.DTOs;
 using API.Utilities.Extensions;
@@ -13,15 +14,15 @@ public partial class GamesController
     /// Creates an admin note for a game
     /// </summary>
     /// <param name="id">Game id</param>
-    /// <response code="404">If a game matching the given id does not exist</response>
-    /// <response code="400">If the authorized user does not exist</response>
+    /// <response code="404">A game matching the given id does not exist</response>
+    /// <response code="400">The authorized user does not exist</response>
     /// <response code="200">Returns the created admin note</response>
     [HttpPost("{id:int}/notes")]
-    [Authorize(Roles = $"{OtrClaims.Roles.Admin}")]
+    [Authorize(Roles = OtrClaims.Roles.Admin)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
     [ProducesResponseType<AdminNoteDTO>(StatusCodes.Status200OK)]
-    public async Task<IActionResult> CreateAdminNoteAsync(int id, [FromBody] string note)
+    public async Task<IActionResult> CreateAdminNoteAsync(int id, [FromBody][Required] string note)
     {
         if (!await gamesService.ExistsAsync(id))
         {
@@ -35,42 +36,24 @@ public partial class GamesController
     }
 
     /// <summary>
-    /// List all admin notes from a game
-    /// </summary>
-    /// <param name="id">Game id</param>
-    /// <response code="404">If a game matching the given id does not exist</response>
-    /// <response code="200">Returns all admin notes from a game</response>
-    [HttpGet("{id:int}/notes")]
-    [Authorize(Roles = $"{OtrClaims.Roles.Admin}")]
-    [ProducesResponseType(StatusCodes.Status404NotFound)]
-    [ProducesResponseType<IEnumerable<AdminNoteDTO>>(StatusCodes.Status200OK)]
-    public async Task<IActionResult> ListAdminNotesAsync(int id)
-    {
-        if (!await gamesService.ExistsAsync(id))
-        {
-            return NotFound();
-        }
-
-        return Ok(await adminNoteService.ListAsync<GameAdminNote>(id));
-    }
-
-    /// <summary>
     /// Updates an admin note for a game
     /// </summary>
     /// <param name="id">Game id</param>
     /// <param name="noteId">Admin note id</param>
     /// <response code="404">
-    /// If a game matching the given id does not exist.
-    /// If an admin note matching the given noteId does not exist
+    /// A game matching the given id does not exist
+    /// or an admin note matching the given noteId does not exist
     /// </response>
-    /// <response code="403">If the requester did not create the admin note</response>
     /// <response code="200">Returns the updated admin note</response>
     [HttpPatch("{id:int}/notes/{noteId:int}")]
-    [Authorize(Roles = $"{OtrClaims.Roles.Admin}")]
+    [Authorize(Roles = OtrClaims.Roles.Admin)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
-    [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [ProducesResponseType<AdminNoteDTO>(StatusCodes.Status200OK)]
-    public async Task<IActionResult> UpdateAdminNoteAsync(int id, int noteId, [FromBody] string note)
+    public async Task<IActionResult> UpdateAdminNoteAsync(
+        int id,
+        int noteId,
+        [FromBody][Required] string note
+    )
     {
         if (!await gamesService.ExistsAsync(id))
         {
@@ -97,12 +80,12 @@ public partial class GamesController
     /// <param name="id">Game id</param>
     /// <param name="noteId">Admin note id</param>
     /// <response code="404">
-    /// If a game matching the given id does not exist.
-    /// If an admin note matching the given noteId does not exist
+    /// A game matching the given id does not exist
+    /// or an admin note matching the given noteId does not exist
     /// </response>
     /// <response code="200">Returns the updated admin note</response>
     [HttpDelete("{id:int}/notes/{noteId:int}")]
-    [Authorize(Roles = $"{OtrClaims.Roles.Admin}")]
+    [Authorize(Roles = OtrClaims.Roles.Admin)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [ProducesResponseType<AdminNoteDTO>(StatusCodes.Status200OK)]
