+++ API/Controllers/GamesController.cs
@@ -3,6 +3,7 @@ using API.DTOs;
 using API.Services.Interfaces;
 using API.Utilities.Extensions;
 using Asp.Versioning;
+using Database.Entities;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.JsonPatch;
 using Microsoft.AspNetCore.Mvc;
@@ -17,17 +18,16 @@ public partial class GamesController(IGamesService gamesService, IAdminNoteServi
     /// <summary>
     /// Amend game data
     /// </summary>
-    /// <param name="id">The game id</param>
+    /// <param name="id">Game id</param>
     /// <param name="patch">JsonPatch data</param>
-    /// <response code="404">If the provided id does not belong to a game</response>
-    /// <response code="400">If JsonPatch data is malformed</response>
-    /// <response code="200">Returns the patched game</response>
+    /// <response code="404">A game matching the given id does not exist</response>
+    /// <response code="400">The JsonPatch data is malformed</response>
+    /// <response code="200">Returns the updated game</response>
+    [HttpPatch("{id:int}")]
     [Authorize(Roles = OtrClaims.Roles.Admin)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    [ProducesResponseType<string>(StatusCodes.Status400BadRequest)]
     [ProducesResponseType<GameDTO>(StatusCodes.Status200OK)]
-    [HttpPatch("{id:int}")]
     public async Task<IActionResult> UpdateAsync(int id, [FromBody] JsonPatchDocument<GameDTO> patch)
     {
         // Ensure target tournament exists
@@ -38,7 +38,7 @@ public partial class GamesController(IGamesService gamesService, IAdminNoteServi
         }
 
         // Ensure request is only attempting to perform a replace operation.
-        if (!patch.IsReplaceOnly())
+        if (patch.Operations.Count == 0 || !patch.IsReplaceOnly())
         {
             return BadRequest();
         }
@@ -59,7 +59,7 @@ public partial class GamesController(IGamesService gamesService, IAdminNoteServi
     /// Delete a game
     /// </summary>
     /// <param name="id">Game id</param>
-    /// <response code="404">The game does not exist</response>
+    /// <response code="404">A game matching the given id does not exist</response>
     /// <response code="204">The game was deleted successfully</response>
     [HttpDelete("{id:int}")]
     [Authorize(Roles = OtrClaims.Roles.Admin)]
@@ -74,5 +74,25 @@ public partial class GamesController(IGamesService gamesService, IAdminNoteServi
 
         await gamesService.DeleteAsync(id);
         return NoContent();
+    }
+
+    /// <summary>
+    /// List all admin notes from a game
+    /// </summary>
+    /// <param name="id">Game id</param>
+    /// <response code="404">A game matching the given id does not exist</response>
+    /// <response code="200">Returns all admin notes from a game</response>
+    [HttpGet("{id:int}/notes")]
+    [Authorize(Roles = $"{OtrClaims.Roles.User}, {OtrClaims.Roles.Client}")]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType<IEnumerable<AdminNoteDTO>>(StatusCodes.Status200OK)]
+    public async Task<IActionResult> ListAdminNotesAsync(int id)
+    {
+        if (!await gamesService.ExistsAsync(id))
+        {
+            return NotFound();
+        }
+
+        return Ok(await adminNoteService.ListAsync<GameAdminNote>(id));
     }
 }
